split api handling
