+++ portfolio/src/main/java/com/google/sps/servlets/DeleteCommentsServlet.java
@@ -1,0 +1,19 @@
+package com.google.sps.servlets;
+
+import java.io.IOException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+/** Servlet that handles deletion of all comments. */
+@WebServlet("/delete-comments")
+public class DeleteCommentsServlet extends HttpServlet {
+
+  @Override
+  public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    CommentsDatastore commentsDatastore = new CommentsDatastore();
+    commentsDatastore.deleteComments();
+  }
+
+}
